+++ autohooks/cli/activate.py
@@ -15,8 +15,6 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import sys
-
 from argparse import Namespace
 
 from autohooks.config import (
@@ -24,32 +22,37 @@ from autohooks.config import (
     get_pyproject_toml_path,
     AUTOHOOKS_SECTION,
 )
-from autohooks.install import (
-    install_pre_commit_hook,
-    get_pre_commit_hook_path,
-    get_autohooks_pre_commit_hook,
-)
-from autohooks.setting import Mode
+from autohooks.hooks import PreCommitHook
+from autohooks.settings import Mode
+from autohooks.terminal import ok, warning, info
 
 
 def install_hooks(args: Namespace) -> None:
-    pre_commit_hook_path = get_pre_commit_hook_path()
+    pre_commit_hook = PreCommitHook()
     pyproject_toml = get_pyproject_toml_path()
     config = load_config_from_pyproject_toml(pyproject_toml)
 
-    if pre_commit_hook_path.exists() and not args.force:
-        print(
-            'pre-commit hook is already installed at {}. --force to '
-            'override.'.format(str(pre_commit_hook_path))
+    if pre_commit_hook.exists() and not args.force:
+        ok(
+            'pre-commit hook is already installed at {}.'.format(
+                str(pre_commit_hook)
+            )
+        )
+        info(
+            "Run 'autohooks activate --force' to override the current "
+            "installed pre-commit hook."
+        )
+        info(
+            "Run 'autohooks check' to validate the current status of "
+            "the installed pre-commit hook."
         )
     else:
         if not config.is_autohooks_enabled():
-            print(
+            warning(
                 'Warning: autohooks is not enabled in your {} file. Please add '
                 'a "{}" section. Run autohooks check for more details.'.format(
                     str(pyproject_toml), AUTOHOOKS_SECTION
-                ),
-                file=sys.stderr,
+                )
             )
 
         if args.mode:
@@ -57,11 +60,10 @@ def install_hooks(args: Namespace) -> None:
         else:
             mode = config.get_mode()
 
-        autohooks_pre_commit_hook = get_autohooks_pre_commit_hook(mode)
-        install_pre_commit_hook(autohooks_pre_commit_hook, pre_commit_hook_path)
+        pre_commit_hook.write(mode=mode)
 
-        print(
+        ok(
             'pre-commit hook installed at {} using {} mode'.format(
-                str(pre_commit_hook_path), str(mode.get_effective_mode())
+                str(pre_commit_hook), str(mode.get_effective_mode())
             )
         )